requirements.txt
numpy~=1.24.4
librosa~=0.10.2.post1
matplotlib~=3.7.5
scipy~=1.10.1
umap~=0.1.1
torchmetrics~=1.4.3
visdom~=0.2.4
seaborn~=0.13.2
webrtcvad~=2.0.10
scikit-learn~=1.3.2
multiprocess~=0.70.16
tqdm~=4.66.5
jiwer~=3.0.5
utmos~=1.1.10
speechmos~=0.0.1.1
swanlab~=0.6.4
wandb~=0.18.5
torchaudio~=2.4.1
soundfile~=0.12.1
speechbrain~=1.0.1
boltons~=24.0.0
pystoi~=0.4.1
pypesq~=1.2.4
resampy~=0.4.3